fix(bootstrap): try to get the secondary luks label code working properly

+++ scripts/helpers.sh
@@ -46,12 +46,16 @@ function yes_or_no() {
 
 # Need this to avoid some wacky pre-commit hook issues related to if rebuild fails and
 # flake.lock stays staged, which ends up wiping out all changes due to stashing bug
+CLEAN=0
 cleanup_flake_lock() {
+    # If the command succeeds, the justfile will clean up for us
     if [ $? -ne 0 ]; then
-        red "Rebuild failed, cleaning up lock files"
-        git rm --cached -f flake.lock 2>/dev/null || true
-        rm flake.lock 2>/dev/null || true
-        rm "$BUILD_LOG"
+        if [[ $CLEAN -eq 0 ]]; then
+            git rm --cached -f flake.lock 2>/dev/null || true
+            red "Rebuild failed, cleaning up lock files"
+            rm flake.lock 2>/dev/null || true
+            CLEAN=1
+        fi
     fi
 }
